add abstract base parser class and refactor parsers
