+++ cli/scrapers/EventbriteParser.ts
@@ -1,19 +1,10 @@
-import * as cheerio from 'cheerio'
-import fs from 'fs';
-import path from 'path';
-import TurndownService from 'turndown';
-import type { DownloadResult, PageParser, ScrapedEventData } from '../types';
+import type { CheerioAPI } from 'cheerio';
 import { DateTime } from 'luxon';
+import type { ScrapedEventData } from '../types';
+import { BaseParser } from './BaseParser';
 
-export class EventbriteParser implements PageParser {
-  private turndownService!: TurndownService;
-  public scraperOutputDir: string = path.join(process.cwd(), 'scraper-output');
-
-  constructor(){
-    this.initializeTurndownService();
-  }
-
-  async scrapeEventDataFromPage($: cheerio.CheerioAPI, url: string): Promise<ScrapedEventData> {
+export class EventbriteParser extends BaseParser {
+  override async scrapeEventDataFromPage($: CheerioAPI, url: string): Promise<ScrapedEventData> {
     const scrapedData: ScrapedEventData = {};
 
     console.log(`Extracting event title...`);
@@ -200,121 +191,5 @@ export class EventbriteParser implements PageParser {
     if (isNaN(minute)) minute = 0; // Handle cases where minute is not provided
 
     return `${hour.toString().padStart(2, '0')}:${minute.toString().padStart(2, '0')}`;
-  }
-
-  private async downloadImage(imageUrl: string): Promise<DownloadResult> {
-    const result: DownloadResult = {
-      originalUrl: imageUrl,
-      fileName: null,
-      filePath: null
-    };
-    try {
-      // Clean the URL by removing query parameters
-      const cleanHeroImageUrl = imageUrl;
-      console.log(`Hero image found: src=${cleanHeroImageUrl}`);
-
-      // Get file name from cleanHeroImageUrl
-      let fileName = path.basename(cleanHeroImageUrl);
-
-      // Add file extension
-      fileName += '.webp';
-
-      result.fileName = fileName;
-
-      console.log(`Downloading hero image...`);
-      const response = await fetch(cleanHeroImageUrl);
-      if (!response.ok) throw new Error(`Failed to download hero image: ${response.statusText}`);
-
-      const arrayBuffer = await response.arrayBuffer();
-      const buffer = Buffer.from(arrayBuffer);
-
-      const targetFolder = path.join(this.scraperOutputDir, `hero-${Date.now()}`);
-      // Ensure scraper output directory exists
-      if (!fs.existsSync(targetFolder)) {
-        fs.mkdirSync(targetFolder, { recursive: true });
-      }
-      
-      const imagePath = path.join(targetFolder, fileName);
-      fs.writeFileSync(imagePath, buffer);
-      console.log(`Hero image downloaded to ${imagePath}`);
-      result.filePath = imagePath;
-    } catch (error) {
-      console.warn(`Error downloading hero image: ${error}`);
-      result.filePath = imageUrl; // Keep the URL if download fails
-    }
-
-    return result;
-  }
-
-  /**
-   * Convert HTML content to Markdown
-   * @param html The HTML content to convert
-   * @returns The converted Markdown content
-   */
-  private convertHtmlToMarkdown(html: string): string {
-    if (!html || html.trim() === '') {
-      return '';
-    }
-
-    try {
-      console.log('Converting HTML to Markdown...');
-      const markdown = this.turndownService.turndown(html);
-
-      // Clean up the markdown - remove excessive newlines
-      const cleanedMarkdown = markdown
-        .replace(/\n{3,}/g, '\n\n') // Replace 3+ newlines with 2
-        .replace(/^\n+/, '') // Remove leading newlines
-        .replace(/\n+$/, ''); // Remove trailing newlines
-
-      console.log('HTML to Markdown conversion completed.');
-      return cleanedMarkdown;
-    } catch (error) {
-      console.error('Error converting HTML to Markdown:', error);
-      console.warn('Falling back to original HTML content.');
-      return html;
-    }
-  }
-
-  /**
-   * Initialize the Turndown service for HTML to Markdown conversion
-   */
-  private initializeTurndownService(): void {
-    this.turndownService = new TurndownService({
-      headingStyle: 'atx',
-      hr: '---',
-      bulletListMarker: '-',
-      codeBlockStyle: 'fenced',
-      fence: '```',
-      emDelimiter: '_',
-      strongDelimiter: '**',
-      linkStyle: 'inlined',
-      linkReferenceStyle: 'full'
-    });
-
-    // Add custom rules for better conversion
-    this.turndownService.addRule('removeComments', {
-      filter: function (node) {
-        return node.nodeType === 8; // Comment node
-      },
-      replacement: function () {
-        return '';
-      }
-    });
-
-    // Remove script and style tags
-    this.turndownService.addRule('removeScriptsAndStyles', {
-      filter: ['script', 'style'],
-      replacement: function () {
-        return '';
-      }
-    });
-
-    // Handle div tags as block elements
-    this.turndownService.addRule('divAsBlock', {
-      filter: 'div',
-      replacement: function (content) {
-        return content ? '\n\n' + content + '\n\n' : '';
-      }
-    });
   }
 }
